clean floders and organize css files
