+++ client/src/components/Loginform/Loginform.css
@@ -1,0 +1,28 @@
+button {
+    margin: 5% auto auto auto;
+    padding:10px;
+  }
+
+  
+ form { text-align: center;
+    margin: 5% 30% 5% 30% ;
+    /* background-color: blanchedalmond; */
+    background-color:rgba(255,255,255,0.7);/*背景颜色白色，70%透明度*/
+    padding: 40px;
+  
+  }
+
+
+ body{
+    height: 100%;
+
+    font-family: 'Righteous', cursive;
+    /* font-family: 'Literata', serif; */
+    
+    display: flex;
+    min-height: 100vh;
+    flex-direction: column;
+    text-align: center;
+  }
+  
+
