+++ client/src/pages/Carousel.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 // import Footer from "../components/Footer";
 // import NavBar from "../components/NavBar";
-import Carousel from "../../components/Carousel";
+import Carousel from "../components/Carousel";
 // import "./Carousel.css"
 
 class CardSave extends Component {
