+++ client/src/pages/IdiomSearchPage.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
-import Footer from "../../components/Footer";
-import NavBar from "../../components/NavBar";
-import { Input, FormBtn } from "../../components/Loginform/index";
+import Footer from "../components/Footer";
+import NavBar from "../components/NavBar";
+import { Input, FormBtn } from "../components/Loginform/index";
 
 class IdiomSearchPage extends Component {
   render() {
