--- client/src/pages/IdiomSearchPage/IdiomSearchPage.css
@@ -1,17 +1,0 @@
-body{
-    background:
-    url('004.jpg') no-repeat center center fixed;
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size:cover;
-    
-    background-size:cover;
-    
-    /* font-family: 'Righteous', cursive; */
-    font-family: 'Literata', serif;
-    
-    display: flex;
-    min-height: 100vh;
-    flex-direction: column;
-    text-align: center;
-   }
