+++ client/src/pages/StartPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import BananaLogo from "../../components/BananaLogo";
+import BananaLogo from "../components/BananaLogo";
 
 class StartPage extends Component {
   render() {
